+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # Changelog
 
-## Upcoming
+## v0.3.0
 
-- PDF tagging can be toggled via `tagged` in `[tool.oicana.export.pdf]` (defaults to `true`)
 - previous `compile` methods in all integrations are now called `export`
 - add new `compile` methods that return compiled documents, which can be
   - exported to any format
@@ -10,6 +9,7 @@
 - templates and compiled documents offer `exportPdf`/`exportPng`/`exportSvg` convenience
   methods in all integrations
 - compiled documents expose the compilation `warnings` in all integrations
+- PDF tagging can be toggled via `tagged` in `[tool.oicana.export.pdf]` (defaults to `true`)
 
 
 ## v0.2.0